Implement camera movement and rotation, start adding mouse-related functions

+++ src/vec3.h
@@ -14,28 +14,28 @@ static inline i16 dot_product(Vec3 a, Vec3 b) {
 }
 
 void rotate_x(Vec3* point, i16 angle) {
-    i16 cosA = cos_degrees(angle);
-    i16 sinA = sin_degrees(angle);
-    i16 y = point->y << 8;
-    i16 z = point->z << 8;
+    i32 cosA = cos_degrees(angle);
+    i32 sinA = sin_degrees(angle);
+    i32 y = point->y << 8;
+    i32 z = point->z << 8;
     
     point->y = (((y * cosA) >> 8) - ((z * sinA) >> 8)) >> 8;
     point->z = (((y * sinA) >> 8) + ((z * cosA) >> 8)) >> 8;
 }
 void rotate_y(Vec3* point, i16 angle) {
-    i16 cosA = cos_degrees(angle);
-    i16 sinA = sin_degrees(angle);
-    i16 x = point->x << 8;
-    i16 z = point->z << 8;
+    i32 cosA = cos_degrees(angle);
+    i32 sinA = sin_degrees(angle);
+    i32 x = point->x << 8;
+    i32 z = point->z << 8;
     
     point->x = (((x * cosA) >> 8) + ((z * sinA) >> 8)) >> 8;
     point->z = (((-x * sinA) >> 8) + ((z * cosA) >> 8)) >> 8;
 }
 void rotate_z(Vec3* point, i16 angle) {
-    i16 cosA = cos_degrees(angle);
-    i16 sinA = sin_degrees(angle);
-    i16 x = point->x << 8;
-    i16 y = point->y << 8;
+    i32 cosA = cos_degrees(angle);
+    i32 sinA = sin_degrees(angle);
+    i32 x = point->x << 8;
+    i32 y = point->y << 8;
     
     point->x = (((x * cosA) >> 8) - ((y * sinA) >> 8)) >> 8;
     point->y = (((x * sinA) >> 8) + ((y * cosA) >> 8)) >> 8;
